Move saving user to service object
